Modules Api

+++ app/Http/Controllers/ModuleApiController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Module;
+use JWTAuth;
 use Illuminate\Http\Request;
 
 class ModuleApiController extends Controller
@@ -12,9 +13,14 @@ class ModuleApiController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        return Module::all();
+        $user = JWTAuth::toUser($request->token);
+        if ($user->hasRole('admin') || $user->hasRole('subscriber')) {
+            return Module::all();
+        } else {
+            return response()->json(['result' => abort(403, 'Unauthorized action.')]);
+        }
     }
 
     /**
@@ -35,9 +41,39 @@ class ModuleApiController extends Controller
      */
     public function store(Request $request)
     {
-        $module = Module::create($request->all());
-
-        return response()->json($module, 201);
+        $validator = \Validator::make(
+                
+            [
+                'title' => $request->title,
+                'slug' => $request->slug,
+                'content' => $request->content
+            ],
+                
+            [
+                'title' => 'sometimes|string|max:255',
+                'slug' => 'sometimes|string|max:255',
+                'content' => 'sometimes|string|max:255',
+            ]
+        );
+        
+        $user = JWTAuth::toUser($request->token);
+        if ($user->hasRole('admin')) {
+            if ($validator->fails()){
+                $result = ['result' => 'Failed',
+                'message' => $validator->errors()];
+                return \Response::json($result)->setStatusCode(400, 'Fail');
+            } else {
+                $module = new Module();
+                $module->title = $request->title;
+                $module->slug = $request->slug;
+                $module->content = $request->content;
+                $module->user_id = $user->id;
+                $module->save();
+                return response()->json($module, 201);
+            }
+        } else {
+            return response()->json(['result' => abort(403, 'Unauthorized action.')]);
+        }
     }
 
     /**
@@ -46,9 +82,14 @@ class ModuleApiController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show(Module $module)
+    public function show(Request $request, Module $module)
     {
-        return $module;
+        $user = JWTAuth::toUser($request->token);
+        if($user->hasRole('admin') || $user->hasRole('subscriber')){
+            return $module;
+        } else {
+            return response()->json(['result' => abort(403, 'Unauthorized action.')]);
+        }
     }
 
     /**
@@ -71,9 +112,13 @@ class ModuleApiController extends Controller
      */
     public function update(Request $request, Module $module)
     {
-        $module->update($request->all());
-
-        return response()->json($module, 200);
+        $user = JWTAuth::toUser($request->token);
+        if($user->hasRole('admin')){
+            $module->update($request->all());
+            return response()->json($module, 200);
+        } else {
+            return response()->json(['result' => abort(403, 'Unauthorized action.')]);
+        }
     }
 
     /**
@@ -82,10 +127,14 @@ class ModuleApiController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Module $module)
+    public function destroy(Request $request, Module $module)
     {
-        $module->delete();
-
-        return response()->json(null, 204);
+        $user = JWTAuth::toUser($request->token);
+        if ($user->hasRole('admin')) {
+            $module->delete();
+            return response()->json(null, 204);
+        } else {
+            return response()->json(['result' => abort(403, 'Unauthorized action.')]);
+        }
     }
 }
